Reuse create context
